fix: split out handlers
