+++ handlers/fees.go
@@ -1,25 +1,16 @@
-package main
+package handlers
 
 import (
 	"math/big"
 
 	evmClient "github.com/coinbase/rosetta-geth-sdk/client"
 	RosettaTypes "github.com/coinbase/rosetta-sdk-go/types"
-	"github.com/ethereum/go-ethereum/common"
 	EthTypes "github.com/ethereum/go-ethereum/core/types"
 
 	sdkTypes "github.com/coinbase/rosetta-geth-sdk/types"
 )
 
-var (
-	baseFeeVault = common.HexToAddress("0x4200000000000000000000000000000000000019")
-	l1FeeVault   = common.HexToAddress("0x420000000000000000000000000000000000001a")
-
-	MintOpType = "MINT"
-	BurnOpType = "BURN"
-)
-
-// FeeOps returns the fee operations for a given transaction
+// FeeOps returns the fee operations for a given transaction.
 func FeeOps(tx *evmClient.LoadedTransaction) ([]*RosettaTypes.Operation, error) {
 	if tx.Transaction.IsDepositTx() {
 		return nil, nil
@@ -89,7 +80,7 @@ func FeeOps(tx *evmClient.LoadedTransaction) ([]*RosettaTypes.Operation, error) 
 			Type:   sdkTypes.FeeOpType,
 			Status: RosettaTypes.String(sdkTypes.SuccessStatus),
 			Account: &RosettaTypes.AccountIdentifier{
-				Address: baseFeeVault.Hex(),
+				Address: BaseFeeVault.Hex(),
 			},
 			// Note: The basefee is not actually burned on L2
 			Amount: evmClient.Amount(tx.FeeBurned, sdkTypes.Currency),
@@ -107,50 +98,11 @@ func FeeOps(tx *evmClient.LoadedTransaction) ([]*RosettaTypes.Operation, error) 
 			Type:   sdkTypes.FeeOpType,
 			Status: RosettaTypes.String(sdkTypes.SuccessStatus),
 			Account: &RosettaTypes.AccountIdentifier{
-				Address: l1FeeVault.Hex(),
+				Address: L1FeeVault.Hex(),
 			},
 			Amount: evmClient.Amount(receipt.L1Fee, sdkTypes.Currency),
 		},
 	}
 
 	return ops, nil
-}
-
-func MintOps(tx *evmClient.LoadedTransaction, startIndex int) []*RosettaTypes.Operation {
-	if tx.Transaction.Mint() == nil {
-		return nil
-	}
-	return []*RosettaTypes.Operation{
-		{
-			OperationIdentifier: &RosettaTypes.OperationIdentifier{
-				Index: int64(startIndex),
-			},
-			Type:   MintOpType,
-			Status: RosettaTypes.String(sdkTypes.SuccessStatus),
-			Account: &RosettaTypes.AccountIdentifier{
-				Address: tx.From.String(),
-			},
-			Amount: evmClient.Amount(tx.Transaction.Mint(), sdkTypes.Currency),
-		},
-	}
-}
-
-// FIXME: this is wrong
-func BurnOps(tx *evmClient.LoadedTransaction, startIndex int) []*RosettaTypes.Operation {
-	if *tx.Transaction.To() != common.HexToAddress("0xdeaddeaddeaddeaddeaddeaddeaddeaddeaddead") {
-		return nil
-	}
-	return []*RosettaTypes.Operation{
-		{
-			OperationIdentifier: &RosettaTypes.OperationIdentifier{
-				Index: int64(startIndex),
-			},
-			Type:   BurnOpType,
-			Status: RosettaTypes.String(sdkTypes.SuccessStatus),
-			Account: &RosettaTypes.AccountIdentifier{
-				Address: tx.From.String(),
-			},
-			Amount: evmClient.Amount(tx.Transaction.Value(), sdkTypes.Currency),
-		},
-	}
 }
